pyproject.toml
[project]
name = "template-backend"
version = "0.1.0"
description = "Backend service scaffold"
authors = [
    {name = "Sal-sal-sal",email = "ccgtru@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastapi (>=0.123.8,<0.124.0)",
    "uvicorn[standard] (>=0.38.0,<0.39.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "alembic (>=1.17.2,<2.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "redis (>=5,<6)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "email-validator (>=2.3.0,<3.0.0)",
    "bcrypt (==4.1.2)"
]

[dependency-groups]
dev = [
    "ruff (>=0.14.8,<0.15.0)",
    "pytest (>=9.0.1,<10.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false